\name{CompositionalRF-package}
\alias{CompositionalRF-package}
\docType{package}
\title{
Multivariate Random Forests with Compositional Responses
}

\description{
Multivariate random forest with compositional response variables and continuous predictor variables.
The data are first transformed using the additive log-ratio transformation and then the multivariate random forest of Rahman R., Otridge J. and Pal R. (2017), <doi:10.1093/bioinformatics/btw765>, is applied.
}

\details{
\tabular{ll}{
Package: \tab CompositionalRF\cr
Type: \tab Package\cr
Version: \tab 1.7 \cr
Date: \tab 2026-09-10\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>
}

\references{
Rahman R., Otridge J. and Pal R. (2017). IntegratedMRF: random forest-based framework for
integrating prediction from different data types. Bioinformatics, 33(9): 1407--1410.

Segal M. and Xiao Y. (2011). Multivariate random forests.
Wiley Interdisciplinary Reviews: Data Mining and Knowledge Discovery, 1(1): 80--87.

Tsagris M.T., Preston S. and Wood A.T.A. (2011). A data-based power transformation
for compositional data. In Proceedings of the 4th Compositional Data Analysis
Workshop, Girona, Spain.
https://arxiv.org/pdf/1106.1451.pdf

Alenazi A. (2023). A review of compositional data analysis and recent advances.
Communications in Statistics--Theory and Methods, 52(16): 5535--5567.

Friedman Jerome, Trevor Hastie and Robert Tibshirani (2009).
The elements of statistical learning, 2nd edition. Springer, Berlin.
}

%\note{
%}


